+++ empresa/src/main/java/RPU/NET/project/empresa/service/IMovimientoDineroService.java
@@ -7,11 +7,11 @@ import RPU.NET.project.empresa.entity.MovimientoDinero;
 import java.util.List;
 
 public interface IMovimientoDineroService {
-    public MovimientoDinero findById(int id);
+    public MovimientoDinero findById(long id);
     public List<MovimientoDinero> findByAll();
     public MovimientoDinero createMovimientoDinero(MovimientoDinero movimientoDinero);
-    public MovimientoDinero updateMovimientoDinero(int id, MovimientoDinero movimientoDinero);
-    public void deletemovimientoDinero(int id);
+    public MovimientoDinero updateMovimientoDinero(long id, MovimientoDinero movimientoDinero);
+    public void deletemovimientoDinero(long id);
 
 
 }
